Move initialization logic to layout
Prevents unnecessary refresh on screen change
Also helps make gallery loading feel instant from initial (conversion) page
