+++ src/lib/stores.ts
@@ -6,6 +6,8 @@ import type { MediaFormat } from '$lib/media';
 export const mediaType = writable<MediaType | null>(null);
 export const previewUrl = writable<string | null>(null);
 export const showMediaPreviewModal = writable(false);
+export const loadingStoredMedia = writable(true);
+export const isDraggingOver = writable(false);
 export const advancedMode = writable(false);
 export const loading = writable(false);
 export const isFFmpegInitialized = writable(false);
